+++ src/main/java/shop/wannab/couponservice/domain/couponpolicy/dto/CategoryDto.java
@@ -1,4 +1,4 @@
-package shop.wannab.couponservice.domain.dto;
+package shop.wannab.couponservice.domain.couponpolicy.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,7 +9,6 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class OrderDto {
-    private long orderId;
-    private long orderBookId;
+public class CategoryDto {
+    private long categoryId;
 }
