+++ src/main/java/shop/wannab/couponservice/domain/couponpolicy/dto/CreateCouponPolicyDto.java
@@ -1,4 +1,4 @@
-package shop.wannab.couponservice.domain.dto;
+package shop.wannab.couponservice.domain.couponpolicy.dto;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalDate;
@@ -9,12 +9,12 @@ import lombok.Setter;
 @Setter
 public class CreateCouponPolicyDto {
     //쿠폰 등록 모달 참조
+    @NotNull
+    private String couponType;
+
     private long bookId;
 
     private long categoryId;
-
-    @NotNull
-    private String couponType;
 
     @NotNull
     private String name;
