+++ src/main/java/shop/wannab/couponservice/domain/couponpolicy/dto/OrderDto.java
@@ -1,4 +1,4 @@
-package shop.wannab.couponservice.domain.dto;
+package shop.wannab.couponservice.domain.couponpolicy.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -9,6 +9,7 @@ import lombok.Setter;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-public class CategoryDto {
-    private long categoryId;
+public class OrderDto {
+    private long orderId;
+    private long orderBookId;
 }
