+++ src/main/java/shop/wannab/couponservice/domain/couponpolicy/dto/UpdateCouponPolicyDto.java
@@ -1,4 +1,4 @@
-package shop.wannab.couponservice.domain.dto;
+package shop.wannab.couponservice.domain.couponpolicy.dto;
 
 import jakarta.validation.constraints.Min;
 import java.time.LocalDate;
