+++ src/main/java/shop/wannab/couponservice/service/CouponPolicyService.java
@@ -9,10 +9,10 @@ import org.springframework.transaction.annotation.Transactional;
 import shop.wannab.couponservice.domain.CouponPolicy;
 import shop.wannab.couponservice.domain.PolicyTargetBook;
 import shop.wannab.couponservice.domain.PolicyTargetCategory;
-import shop.wannab.couponservice.domain.dto.CouponPolicyDetailResponseDto;
-import shop.wannab.couponservice.domain.dto.CouponPolicyResponseDto;
-import shop.wannab.couponservice.domain.dto.CreateCouponPolicyDto;
-import shop.wannab.couponservice.domain.dto.UpdateCouponPolicyDto;
+import shop.wannab.couponservice.domain.couponpolicy.dto.CouponPolicyDetailResponseDto;
+import shop.wannab.couponservice.domain.couponpolicy.dto.CouponPolicyResponseDto;
+import shop.wannab.couponservice.domain.couponpolicy.dto.CreateCouponPolicyDto;
+import shop.wannab.couponservice.domain.couponpolicy.dto.UpdateCouponPolicyDto;
 import shop.wannab.couponservice.domain.enums.DiscountType;
 import shop.wannab.couponservice.domain.enums.PolicyRule;
 import shop.wannab.couponservice.domain.enums.PolicyStatus;
